requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
psycopg2==2.9.10
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1